Split Template functions and method registry
